--- model/src/main/java/com/epam/esm/domain/validation/PatchDto.java
@@ -1,4 +1,0 @@
-package com.epam.esm.domain.validation;
-
-public @interface PatchDto {
-}
