+++ model/src/main/java/com/epam/esm/model/AbstractEntity.java
@@ -1,10 +1,9 @@
-package com.epam.esm.domain;
+package com.epam.esm.model;
 
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
 import javax.persistence.Id;
 import javax.persistence.MappedSuperclass;
-import java.util.Objects;
 
 @MappedSuperclass
 public abstract class AbstractEntity {
