+++ model/src/main/java/com/epam/esm/model/dto/GiftCertificateDto.java
@@ -1,10 +1,9 @@
-package com.epam.esm.domain.dto;
+package com.epam.esm.model.dto;
 
-import com.epam.esm.domain.Tag;
-import com.epam.esm.domain.validation.DtoTag;
-import com.epam.esm.domain.validation.PatchDto;
-import com.epam.esm.domain.validation.SaveDto;
-import com.epam.esm.domain.validation.ValidationErrorCode;
+import com.epam.esm.model.validation.DtoTag;
+import com.epam.esm.model.validation.PatchDto;
+import com.epam.esm.model.validation.SaveDto;
+import com.epam.esm.model.validation.ValidationErrorCode;
 import com.fasterxml.jackson.annotation.JsonFormat;
 
 import javax.validation.constraints.NotNull;
