+++ model/src/main/java/com/epam/esm/model/dto/UserDto.java
@@ -1,4 +1,4 @@
-package com.epam.esm.domain.dto;
+package com.epam.esm.model.dto;
 
 public class UserDto {
     private Long id;
