+++ model/src/main/java/com/epam/esm/model/validation/PatchDto.java
@@ -1,0 +1,4 @@
+package com.epam.esm.model.validation;
+
+public @interface PatchDto {
+}
