+++ persistence/src/main/java/com/epam/esm/persistence/repository/CrudRepository.java
@@ -1,6 +1,6 @@
 package com.epam.esm.persistence.repository;
 
-import com.epam.esm.domain.AbstractEntity;
+import com.epam.esm.model.AbstractEntity;
 
 import java.util.List;
