+++ persistence/src/main/java/com/epam/esm/persistence/repository/OrderRepository.java
@@ -1,6 +1,6 @@
 package com.epam.esm.persistence.repository;
 
-import com.epam.esm.domain.Order;
+import com.epam.esm.model.Order;
 
 public interface OrderRepository extends CrudRepository<Order> {
