+++ persistence/src/main/java/com/epam/esm/persistence/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.epam.esm.persistence.repository;
 
-import com.epam.esm.domain.User;
+import com.epam.esm.model.User;
 
 public interface UserRepository extends CrudRepository<User> {
 }
