+++ service/src/main/java/com/epam/esm/service/GiftCertificateService.java
@@ -1,7 +1,7 @@
 package com.epam.esm.service;
 
-import com.epam.esm.domain.dto.GiftCertificateDto;
-import com.epam.esm.domain.dto.TagDto;
+import com.epam.esm.model.dto.GiftCertificateDto;
+import com.epam.esm.model.dto.TagDto;
 
 import java.util.List;
 import java.util.Optional;
