+++ service/src/main/java/com/epam/esm/service/impl/UserServiceImpl.java
@@ -1,10 +1,10 @@
 package com.epam.esm.service.impl;
 
-import com.epam.esm.domain.Order;
-import com.epam.esm.domain.User;
-import com.epam.esm.domain.dto.OrderDto;
-import com.epam.esm.domain.dto.UserDto;
-import com.epam.esm.domain.dto.serialization.DtoSerializer;
+import com.epam.esm.model.Order;
+import com.epam.esm.model.User;
+import com.epam.esm.model.dto.OrderDto;
+import com.epam.esm.model.dto.UserDto;
+import com.epam.esm.model.dto.serialization.DtoSerializer;
 import com.epam.esm.exception.ErrorCode;
 import com.epam.esm.exception.ServiceException;
 import com.epam.esm.persistence.repository.UserRepository;
