+++ service/src/test/java/com/epam/esm/service/impl/UserServiceImplTest.java
@@ -1,13 +1,13 @@
 package com.epam.esm.service.impl;
 
-import com.epam.esm.domain.GiftCertificate;
-import com.epam.esm.domain.Order;
-import com.epam.esm.domain.User;
-import com.epam.esm.domain.dto.GiftCertificateDto;
-import com.epam.esm.domain.dto.OrderDto;
-import com.epam.esm.domain.dto.UserDto;
-import com.epam.esm.domain.dto.serialization.OrderDtoSerializer;
-import com.epam.esm.domain.dto.serialization.UserDtoSerializer;
+import com.epam.esm.model.GiftCertificate;
+import com.epam.esm.model.Order;
+import com.epam.esm.model.User;
+import com.epam.esm.model.dto.GiftCertificateDto;
+import com.epam.esm.model.dto.OrderDto;
+import com.epam.esm.model.dto.UserDto;
+import com.epam.esm.model.dto.serialization.OrderDtoSerializer;
+import com.epam.esm.model.dto.serialization.UserDtoSerializer;
 import com.epam.esm.exception.ServiceException;
 import com.epam.esm.persistence.repository.UserRepository;
 import com.epam.esm.service.UserService;
