+++ web/src/main/java/com/epam/esm/web/controller/OrdersController.java
@@ -1,6 +1,6 @@
 package com.epam.esm.web.controller;
 
-import com.epam.esm.domain.dto.OrderDto;
+import com.epam.esm.model.dto.OrderDto;
 import com.epam.esm.service.OrderService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
