+++ web/src/main/java/com/epam/esm/web/controller/UsersController.java
@@ -1,7 +1,7 @@
 package com.epam.esm.web.controller;
 
-import com.epam.esm.domain.dto.OrderDto;
-import com.epam.esm.domain.dto.UserDto;
+import com.epam.esm.model.dto.OrderDto;
+import com.epam.esm.model.dto.UserDto;
 import com.epam.esm.service.UserService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
